@@ -1,7 +1,11 @@
-const fs = require('fs');
-const path = require('path');
-const dotenv = require('dotenv');
-const logger = require('./logger');
+import fs from 'fs';
+import path from 'path';
+import dotenv from 'dotenv';
+import { fileURLToPath } from 'url';
+import logger from './logger.js';
+
+// Get __dirname equivalent in ESM
+const __filename = fileURLToPath(import.meta.url);
 
 /**
  * Load environment variables from .env file
@@ -131,7 +135,7 @@ function getDeploymentConfig() {
   };
 }
 
-module.exports = {
+export {
   getGitHubCredentials,
   getDeploymentConfig,
   loadEnv
